+++ src/styling/GospelDisplay.css
@@ -1,9 +1,27 @@
 .container {
     display: flex;
+    justify-content: space-between;
+    align-items: center; /* Center the sidebar and companion vertically */
+    height: 100vh; /* Full viewport height to vertically center items */
+    padding: 20px;
+    box-sizing: border-box;
+  }
+
+  .sidebar {
+    flex: 1;
+    padding: 10px;
+    gap: 10px;
+    
+    top: 50%;
+    width: 40%;
+    height: 100%;
+    text-align: center;
+    justify-content: center;
+    align-items: center;
   }
   
   .scripture {
-    flex: 2;
+    flex: 3;
     padding: 10px;
     gap: 10px;
 
@@ -13,7 +31,7 @@
   }
 
   .companion {
-    flex: 1;
+    flex: 2;
     padding: 10px;
     gap: 10px;
